requirements.txt
decord==0.6.0
deepspeed==0.16.3
einops==0.8.1
eva_decord==0.6.1
fvcore==0.1.5.post20221221
matplotlib==3.10.1
numpy==1.26.3
openai==1.65.4
opencv_python==4.10.0.84
packaging==24.2
pandas==2.2.3
peft==0.14.0
Pillow==11.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
qwen_vl_utils==0.0.10
Requests==2.32.3
scipy==1.15.2
Shapely==2.0.7
skimage==0.0
termcolor==2.5.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0.dev0